+++ L2_GameServer/java/net/sf/l2j/gameserver/model/L2Manor.java
@@ -104,8 +104,8 @@ public class L2Manor
      * @param seedId
      * @return seed price
      */
-    public int getSeedBuyPrice (int seedId) {
-    	int buyPrice = getSeedBasicPrice(seedId) / 10;
+    public long getSeedBuyPrice (int seedId) {
+    	long buyPrice = getSeedBasicPrice(seedId) / 10;
     	return (buyPrice > 0?buyPrice:1);
     }
