+++ L2_GameServer/java/net/sf/l2j/gameserver/model/L2ManufactureItem.java
@@ -25,10 +25,10 @@ import net.sf.l2j.gameserver.RecipeController;
 public class L2ManufactureItem
 {
     private int _recipeId;
-    private int _cost;
+    private long _cost;
     private boolean _isDwarven;
 
-    public L2ManufactureItem(int recipeId, int cost)
+    public L2ManufactureItem(int recipeId, long cost)
     {
         _recipeId = recipeId;
         _cost = cost;
@@ -41,7 +41,7 @@ public class L2ManufactureItem
         return _recipeId;
     }
 
-    public int getCost()
+    public long getCost()
     {
         return _cost;
     }
